requirements.txt
Django
django-cleanup
django-crispy-forms
djangorestframework
gunicorn
python-decouple
pandas
Pillow

# testing packages
pytest
pytest-django
factory_boy
pytest-xdist